Refactored util fns out
