+++ core/server/lib/members/index.js
@@ -1,6 +1,8 @@
 const {Router, static} = require('express');
 const body = require('body-parser');
 
+const {getData, handleError} = require('./util');
+
 const cookies = require('./cookies');
 const tokens = require('./tokens');
 
@@ -163,37 +165,3 @@ module.exports = function MembersApi({
 
     return httpHandler;
 };
-
-/* http */
-function getData(...props) {
-    return function (req, res, next) {
-        if (!req.body) {
-            res.writeHead(400);
-            return res.end();
-        }
-
-        const data = props.concat('origin').reduce((data, prop) => {
-            if (!data || !req.body[prop]) {
-                return null;
-            }
-            return Object.assign(data, {
-                [prop]: req.body[prop]
-            });
-        }, {});
-
-        if (!data) {
-            res.writeHead(400);
-            return res.end(`Expected {${props.join(', ')}}`);
-        }
-        req.data = data || {};
-        next();
-    };
-}
-
-/* http */
-function handleError(status, res) {
-    return function () {
-        res.writeHead(status);
-        res.end();
-    };
-}
